requirements.txt
alembic==1.13.3
backoff==2.2.1
black==24.10.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
googleapis-common-protos==1.59.1
graypy==2.1.0
greenlet==3.1.1
grpcio==1.67.0
idna==3.10
importlib_metadata==8.4.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.1
mypy-extensions==1.0.0
opentelemetry-api==1.27.0
opentelemetry-distro==0.48b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-exporter-prometheus==0.48b0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asyncio==0.48b0
opentelemetry-instrumentation-aws-lambda==0.43b0
opentelemetry-instrumentation-dbapi==0.48b0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-grpc==0.48b0
opentelemetry-instrumentation-jinja2==0.48b0
opentelemetry-instrumentation-logging==0.48b0
opentelemetry-instrumentation-psycopg2==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-instrumentation-sqlite3==0.48b0
opentelemetry-instrumentation-threading==0.48b0
opentelemetry-instrumentation-urllib==0.48b0
opentelemetry-instrumentation-urllib3==0.48b0
opentelemetry-instrumentation-wsgi==0.48b0
opentelemetry-processor-baggage==0.48b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.0
protobuf==4.25.5
psycopg2==2.9.10
pytest==8.3.3
PyYAML==6.0.2
requests==2.32.3
setuptools==75.2.0
six==1.16.0
SQLAlchemy==2.0.36
thrift==0.21.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wrapt==1.16.0
zipp==3.20.2
